feat: Introduce context provider
