+++ src/Core/System/DependencyInjection/sales_channel.xml
@@ -76,9 +76,14 @@
             <argument type="service" id="tax.repository"/>
             <argument type="service" id="payment_method.repository"/>
             <argument type="service" id="shipping_method.repository"/>
-            <argument type="service" id="Doctrine\DBAL\Connection"/>
             <argument type="service" id="country_state.repository"/>
             <argument type="service" id="currency_country_rounding.repository"/>
+            <argument type="service" id="Shopware\Core\System\SalesChannel\Context\ContextProvider"/>
+        </service>
+
+        <service id="Shopware\Core\System\SalesChannel\Context\ContextProvider">
+            <argument type="service" id="Doctrine\DBAL\Connection"/>
+            <argument type="service" id="event_dispatcher"/>
         </service>
 
         <service id="Shopware\Core\System\SalesChannel\Context\CachedBaseContextFactory" decorates="Shopware\Core\System\SalesChannel\Context\BaseContextFactory">
