+++ src/Core/System/SalesChannel/Context/BaseContextFactory.php
@@ -2,7 +2,6 @@
 
 namespace Shopware\Core\System\SalesChannel\Context;
 
-use Doctrine\DBAL\Connection;
 use Shopware\Core\Checkout\Cart\Delivery\Struct\ShippingLocation;
 use Shopware\Core\Checkout\Cart\Price\Struct\CartPrice;
 use Shopware\Core\Checkout\Customer\Aggregate\CustomerGroup\CustomerGroupCollection;
@@ -10,9 +9,6 @@ use Shopware\Core\Checkout\Payment\PaymentMethodCollection;
 use Shopware\Core\Checkout\Payment\PaymentMethodEntity;
 use Shopware\Core\Checkout\Shipping\ShippingMethodCollection;
 use Shopware\Core\Checkout\Shipping\ShippingMethodEntity;
-use Shopware\Core\Defaults;
-use Shopware\Core\Framework\Api\Context\AdminSalesChannelApiSource;
-use Shopware\Core\Framework\Api\Context\SalesChannelApiSource;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Pricing\CashRoundingConfig;
@@ -60,9 +56,9 @@ class BaseContextFactory extends AbstractBaseContextFactory
         private readonly EntityRepository $taxRepository,
         private readonly EntityRepository $paymentMethodRepository,
         private readonly EntityRepository $shippingMethodRepository,
-        private readonly Connection $connection,
         private readonly EntityRepository $countryStateRepository,
-        private readonly EntityRepository $currencyCountryRepository
+        private readonly EntityRepository $currencyCountryRepository,
+        private readonly ContextProvider $contextProvider,
     ) {
     }
 
@@ -71,7 +67,7 @@ class BaseContextFactory extends AbstractBaseContextFactory
      */
     public function create(string $salesChannelId, array $options = []): BaseContext
     {
-        $context = $this->getContext($salesChannelId, $options);
+        $context = $this->contextProvider->getContext($salesChannelId, $options);
 
         $criteria = new Criteria([$salesChannelId]);
         $criteria->setTitle('base-context-factory::sales-channel');
@@ -216,83 +212,6 @@ class BaseContextFactory extends AbstractBaseContextFactory
     }
 
     /**
-     * @param array<string, mixed> $session
-     */
-    private function getContext(string $salesChannelId, array $session): Context
-    {
-        $sql = '
-        # context-factory::base-context
-
-        SELECT
-          sales_channel.id as sales_channel_id,
-          sales_channel.language_id as sales_channel_default_language_id,
-          sales_channel.currency_id as sales_channel_currency_id,
-          currency.factor as sales_channel_currency_factor,
-          GROUP_CONCAT(LOWER(HEX(sales_channel_language.language_id))) as sales_channel_language_ids
-        FROM sales_channel
-            INNER JOIN currency
-                ON sales_channel.currency_id = currency.id
-            LEFT JOIN sales_channel_language
-                ON sales_channel_language.sales_channel_id = sales_channel.id
-        WHERE sales_channel.id = :id
-        GROUP BY sales_channel.id, sales_channel.language_id, sales_channel.currency_id, currency.factor';
-
-        $data = $this->connection->fetchAssociative($sql, [
-            'id' => Uuid::fromHexToBytes($salesChannelId),
-        ]);
-        if ($data === false) {
-            throw SalesChannelException::noContextData($salesChannelId);
-        }
-
-        if (isset($session[SalesChannelContextService::ORIGINAL_CONTEXT])) {
-            $origin = new AdminSalesChannelApiSource($salesChannelId, $session[SalesChannelContextService::ORIGINAL_CONTEXT]);
-        } else {
-            $origin = new SalesChannelApiSource($salesChannelId);
-        }
-
-        // explode all available languages for the provided sales channel
-        $languageIds = $data['sales_channel_language_ids'] ? explode(',', (string) $data['sales_channel_language_ids']) : [];
-        $languageIds = array_keys(array_flip($languageIds));
-
-        // check which language should be used in the current request (request header set, or context already contains a language - stored in `sales_channel_api_context`)
-        $defaultLanguageId = Uuid::fromBytesToHex($data['sales_channel_default_language_id']);
-
-        $languageChain = $this->buildLanguageChain($session, $defaultLanguageId, $languageIds);
-
-        $versionId = Defaults::LIVE_VERSION;
-        if (isset($session[SalesChannelContextService::VERSION_ID])) {
-            $versionId = $session[SalesChannelContextService::VERSION_ID];
-        }
-
-        return new Context(
-            $origin,
-            [],
-            Uuid::fromBytesToHex($data['sales_channel_currency_id']),
-            $languageChain,
-            $versionId,
-            (float) $data['sales_channel_currency_factor'],
-            true
-        );
-    }
-
-    private function getParentLanguageId(string $languageId): ?string
-    {
-        $data = $this->connection->createQueryBuilder()
-            ->select('LOWER(HEX(language.parent_id))')
-            ->from('language')
-            ->where('language.id = :id')
-            ->setParameter('id', Uuid::fromHexToBytes($languageId))
-            ->executeQuery()
-            ->fetchOne();
-
-        if ($data === false) {
-            throw SalesChannelException::languageNotFound($languageId);
-        }
-
-        return $data;
-    }
-
-    /**
      * @param array<string, mixed> $options
      */
     private function loadShippingLocation(array $options, Context $context, SalesChannelEntity $salesChannel): ShippingLocation
@@ -338,33 +257,6 @@ class BaseContextFactory extends AbstractBaseContextFactory
         }
 
         return ShippingLocation::createFromCountry($country);
-    }
-
-    /**
-     * @param array<string, mixed> $sessionOptions
-     * @param array<string> $availableLanguageIds
-     *
-     * @return non-empty-list<string>
-     */
-    private function buildLanguageChain(array $sessionOptions, string $defaultLanguageId, array $availableLanguageIds): array
-    {
-        $current = $sessionOptions[SalesChannelContextService::LANGUAGE_ID] ?? $defaultLanguageId;
-
-        if (!\is_string($current) || !Uuid::isValid($current)) {
-            throw SalesChannelException::invalidLanguageId();
-        }
-
-        // check provided language is part of the available languages
-        if (!\in_array($current, $availableLanguageIds, true)) {
-            throw SalesChannelException::providedLanguageNotAvailable($current, $availableLanguageIds);
-        }
-
-        if ($current === Defaults::LANGUAGE_SYSTEM) {
-            return [Defaults::LANGUAGE_SYSTEM];
-        }
-
-        // provided language can be a child language
-        return array_values(array_filter([$current, $this->getParentLanguageId($current), Defaults::LANGUAGE_SYSTEM]));
     }
 
     /**
